backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
mysqlclient
pymysql
anthropic
openai
langchain
pdfplumber
pillow
stripe
